requirements.txt
atlassian-python-api
python-dotenv
beautifulsoup4
chromadb
openai
langchain
tiktoken # Required for openai auth
sentence-transformers
fastapi[all]